+++ app/code/local/Stuntcoders/Banner/Block/Adminhtml/Group/Index/Grid.php
@@ -1,0 +1,75 @@
+<?php
+
+class Stuntcoders_Banner_Block_Adminhtml_Group_Index_Grid extends Mage_Adminhtml_Block_Widget_Grid
+{
+    public function __construct()
+    {
+        parent::__construct();
+        $this->setId('bannerGroupGrid');
+        $this->setDefaultSort('id');
+        $this->setDefaultDir('DESC');
+        $this->setSaveParametersInSession(true);
+    }
+
+    /**
+     * @return Mage_Adminhtml_Block_Widget_Grid
+     */
+    protected function _prepareCollection()
+    {
+        $this->setCollection(Mage::getModel('stuntcoders_banner/banner_group')->getCollection());
+        return parent::_prepareCollection();
+    }
+
+    /**
+     * @return Mage_Adminhtml_Block_Widget_Grid
+     */
+    protected function _prepareColumns()
+    {
+        $this->addColumn('id', array(
+            'header' => $this->__('id'),
+            'align' => 'left',
+            'width' => '50px',
+            'index' => 'id',
+        ));
+
+        $this->addColumn('code', array(
+            'header' => $this->__('Code'),
+            'align' => 'left',
+            'index' => 'code',
+        ));
+
+        $this->addColumn('name', array(
+            'header' => $this->__('Name'),
+            'align' => 'left',
+            'index' => 'name',
+        ));
+
+        return parent::_prepareColumns();
+    }
+
+    /**
+     * @return Stuntcoders_Banner_Block_Adminhtml_Group_Index_Grid
+     */
+    protected function _prepareMassaction()
+    {
+        $this->setMassactionIdField('id');
+        $this->getMassactionBlock()->setFormFieldName('groups');
+
+        $this->getMassactionBlock()->addItem('delete', array(
+            'label' => $this->__('Delete'),
+            'url' => $this->getUrl('*/*/massDelete'),
+            'confirm' => $this->__('Are you sure?')
+        ));
+
+        return $this;
+    }
+
+    /**
+     * @param Varien_Object $row
+     * @return string
+     */
+    public function getRowUrl($row)
+    {
+        return $this->getUrl('*/*/new', array('id' => $row->getId()));
+    }
+}