+++ app/code/local/Stuntcoders/Banner/Block/Adminhtml/Index/Grid.php
@@ -1,54 +1,61 @@
 <?php
 
-class Stuntcoders_Banner_Block_Adminhtml_Banner_Grid extends Mage_Adminhtml_Block_Widget_Grid
+class Stuntcoders_Banner_Block_Adminhtml_Index_Grid extends Mage_Adminhtml_Block_Widget_Grid
 {
     public function __construct()
     {
         parent::__construct();
         $this->setId('bannerGrid');
-        $this->setDefaultSort('banner_id');
+        $this->setDefaultSort('id');
         $this->setDefaultDir('DESC');
         $this->setSaveParametersInSession(true);
     }
 
+    /**
+     * @return Mage_Adminhtml_Block_Widget_Grid
+     */
     protected function _prepareCollection()
     {
         $collection = Mage::getModel('stuntcoders_banner/banner')->getCollection();
 
         $this->setCollection($collection);
+
         return parent::_prepareCollection();
     }
 
+    /**
+     * @return Mage_Adminhtml_Block_Widget_Grid
+     */
     protected function _prepareColumns()
     {
-        $this->addColumn('banner_id', array(
-            'header' => Mage::helper('stuntcoders_banner')->__('ID'),
+        $this->addColumn('id', array(
+            'header' => $this->__('ID'),
             'align' => 'left',
             'width' => '50px',
-            'index' => 'banner_id'
+            'index' => 'id'
         ));
 
         $this->addColumn('image', array(
-            'header' => Mage::helper('stuntcoders_banner')->__('Image'),
+            'header' => $this->__('Image'),
             'align' => 'center',
             'index' => 'image',
-            'renderer' => 'stuntcoders_banner/adminhtml_banner_grid_image_renderer',
+            'renderer' => 'stuntcoders_banner/adminhtml_index_grid_image_renderer',
         ));
 
         $this->addColumn('code', array(
-            'header' => Mage::helper('stuntcoders_banner')->__('Code'),
+            'header' => $this->__('Code'),
             'align' => 'left',
             'index' => 'code',
         ));
 
         $this->addColumn('heading', array(
-            'header' => Mage::helper('stuntcoders_banner')->__('Title'),
+            'header' => $this->__('Title'),
             'align' => 'left',
             'index' => 'heading',
         ));
 
         $this->addColumn('group_id', array(
-            'header' => Mage::helper('stuntcoders_banner')->__('Group'),
+            'header' => $this->__('Group'),
             'align' => 'left',
             'index' => 'group_id',
             'type' => 'options',
@@ -56,30 +63,33 @@ class Stuntcoders_Banner_Block_Adminhtml_Banner_Grid extends Mage_Adminhtml_Bloc
         ));
 
         $this->addColumn('url', array(
-            'header' => Mage::helper('stuntcoders_banner')->__('URL'),
+            'header' => $this->__('URL'),
             'align' => 'left',
             'index' => 'url',
         ));
 
         $this->addColumn('text', array(
-            'header' => Mage::helper('stuntcoders_banner')->__('Text'),
-            'align' =>'left',
+            'header' => $this->__('Text'),
+            'align' => 'left',
             'index' => 'text',
-            'renderer' => 'stuntcoders_banner/adminhtml_banner_grid_description_renderer',
+            'renderer' => 'stuntcoders_banner/adminhtml_index_grid_description_renderer',
         ));
 
         return parent::_prepareColumns();
     }
 
+    /**
+     * @return Stuntcoders_Banner_Block_Adminhtml_Index_Grid
+     */
     protected function _prepareMassaction()
     {
-        $this->setMassactionIdField('banner_id');
+        $this->setMassactionIdField('id');
         $this->getMassactionBlock()->setFormFieldName('banners');
 
         $this->getMassactionBlock()->addItem('delete', array(
-            'label' => Mage::helper('stuntcoders_banner')->__('Delete'),
+            'label' => $this->__('Delete'),
             'url' => $this->getUrl('*/*/massDelete'),
-            'confirm' => Mage::helper('stuntcoders_banner')->__('Are you sure?')
+            'confirm' => $this->__('Are you sure?')
         ));
 
         $this->getMassactionBlock()->addItem('group', array(
@@ -90,7 +100,7 @@ class Stuntcoders_Banner_Block_Adminhtml_Banner_Grid extends Mage_Adminhtml_Bloc
                     'name' => 'group',
                     'type' => 'select',
                     'class' => 'required-entry',
-                    'label' => Mage::helper('catalog')->__('Assign to group'),
+                    'label' => $this->__('Assign to group'),
                     'values' => Mage::getModel('stuntcoders_banner/banner_group')->getSelectOptionValues()
                 )
             )
@@ -99,8 +109,12 @@ class Stuntcoders_Banner_Block_Adminhtml_Banner_Grid extends Mage_Adminhtml_Bloc
         return $this;
     }
 
+    /**
+     * @param Varien_Object $row
+     * @return string
+     */
     public function getRowUrl($row)
     {
-        return $this->getUrl('*/*/add', array('id' => $row->getId()));
+        return $this->getUrl('*/*/new', array('id' => $row->getId()));
     }
 }