+++ app/code/local/Stuntcoders/Banner/Block/Adminhtml/Index/Grid/Image/Renderer.php
@@ -1,6 +1,6 @@
 <?php
 
-class Stuntcoders_Banner_Block_Adminhtml_Banner_Grid_Image_Renderer
+class Stuntcoders_Banner_Block_Adminhtml_Index_Grid_Image_Renderer
     extends Mage_Adminhtml_Block_Widget_Grid_Column_Renderer_Abstract
 {
     public function render(Varien_Object $row)