+++ app/code/local/Stuntcoders/Banner/Block/Adminhtml/New/Form.php
@@ -1,13 +1,16 @@
 <?php
 
-class Stuntcoders_Banner_Block_Adminhtml_Banner_Form extends Mage_Adminhtml_Block_Widget_Form
+class Stuntcoders_Banner_Block_Adminhtml_New_Form extends Mage_Adminhtml_Block_Widget_Form
 {
+    /**
+     * @return Mage_Adminhtml_Block_Widget_Form
+     */
     protected function _prepareForm()
     {
         $form = new Varien_Data_Form(
             array(
-                'id' => 'banner_form',
-                'name' => 'banner_form',
+                'id' => 'edit_form',
+                'name' => 'edit_form',
                 'action' => $this->getUrl('*/*/save', array('id' => $this->getRequest()->getParam('id'))),
                 'method' => 'post',
                 'enctype' => 'multipart/form-data'
@@ -31,26 +34,41 @@ class Stuntcoders_Banner_Block_Adminhtml_Banner_Form extends Mage_Adminhtml_Bloc
             'name' => 'code',
         ));
 
+        if (!Mage::app()->isSingleStoreMode()) {
+            $fieldset->addField('store_id', 'multiselect', array(
+                'name' => 'stores[]',
+                'label' => Mage::helper('stuntcoders_banner')->__('Store View'),
+                'title' => Mage::helper('stuntcoders_banner')->__('Store View'),
+                'required' => true,
+                'values' => Mage::getSingleton('adminhtml/system_store')->getStoreValuesForForm(false, true),
+            ));
+        } else {
+            $fieldset->addField('store_id', 'hidden', array(
+                'name' => 'stores[]',
+                'value' => Mage::app()->getStore(true)->getId()
+            ));
+        }
+
         $fieldset->addField('group_id', 'select', array(
             'label' => Mage::helper('stuntcoders_banner')->__('Banner Group'),
             'name' => 'group_id',
             'values' => Mage::getModel('stuntcoders_banner/banner_group')->getSelectOptionValues()
         ));
 
-        $fieldset->addField('heading', 'text', array(
+        $fieldset->addField('title', 'text', array(
             'label' => Mage::helper('stuntcoders_banner')->__('Title'),
-            'name' => 'heading',
+            'name' => 'title',
+        ));
+
+        $fieldset->addField('content', 'textarea', array(
+            'label' => Mage::helper('stuntcoders_banner')->__('Text'),
+            'name' => 'content',
         ));
 
         $fieldset->addField('url', 'text', array(
             'label' => Mage::helper('stuntcoders_banner')->__('Url'),
             'name' => 'url',
             'note' =>  Mage::helper('stuntcoders_banner')->__('On click url'),
-        ));
-
-        $fieldset->addField('text', 'textarea', array(
-            'label' => Mage::helper('stuntcoders_banner')->__('Text'),
-            'name' => 'text',
         ));
 
         $fieldset->addField('image', 'image', array(