+++ app/code/local/Stuntcoders/Banner/controllers/Adminhtml/BannerController.php
@@ -7,7 +7,7 @@ class Stuntcoders_Banner_Adminhtml_BannerController extends Mage_Adminhtml_Contr
         $this->_initAction()->renderLayout();
     }
 
-    public function addAction()
+    public function newAction()
     {
         $this->_initAction()->renderLayout();
     }
@@ -36,10 +36,25 @@ class Stuntcoders_Banner_Adminhtml_BannerController extends Mage_Adminhtml_Contr
                 $banner->save();
 
                 $this->_getSession()->addSuccess($this->__('Item was successfully saved'));
-                $this->_redirect('*/*/add', array('id' => $banner->getId()));
+                $this->_redirect('*/*/new', array('id' => $banner->getId()));
             } catch (Exception $e) {
                 $this->_getSession()->addError($e->getMessage());
-                $this->_redirectReferer('*/*/');
+                $this->_redirectReferer();
+            }
+        }
+    }
+
+    public function deleteAction()
+    {
+        if ($id = $this->getRequest()->getParam('id')) {
+            try {
+                Mage::getModel('stuntcoders_banner/banner')->setId($id)->delete();
+
+                $this->_getSession()->addSuccess($this->__('Banner was successfully deleted'));
+                $this->_redirect('*/*');
+            } catch (Exception $e) {
+                $this->_getSession()->addError($e->getMessage());
+                $this->_redirectReferer();
             }
         }
     }